update to bigdecimal

+++ src/main/java/io/meduse/exchange/MarketDepth.java
@@ -1,5 +1,6 @@
 package io.meduse.exchange;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
@@ -11,36 +12,37 @@ import io.meduse.data.Bucket;
 
 public class MarketDepth {
 
-	List<DepthPoint> bid = null;
-	List<DepthPoint> ask = null;
+  List<DepthPoint> bid = null;
+  List<DepthPoint> ask = null;
 
-	public MarketDepth(OrderBook orderBook) {
-		bid = addPoints(orderBook.getBids(), orderBook.getBidBucket());
-		ask = addPoints(orderBook.getAsks(), orderBook.getAskBucket());
-	}
+  public MarketDepth(OrderBook orderBook) {
+    bid = addPoints(orderBook.getBids(), orderBook.getBidBucket());
+    ask = addPoints(orderBook.getAsks(), orderBook.getAskBucket());
+  }
 
-	private List<DepthPoint> addPoints(NavigableSet<Long> pricePoints, HashMap<Long, Bucket> buckets) {
-		List<DepthPoint> result = new ArrayList<DepthPoint>();
-		for (Long price : pricePoints) {
-			Long volume = buckets.get(price).getVolume();
-			result.add(new DepthPoint(price, volume));
-		}
-		return result;
-	}
+  private List<DepthPoint> addPoints(NavigableSet<BigDecimal> pricePoints,
+      HashMap<BigDecimal, Bucket> buckets) {
+    List<DepthPoint> result = new ArrayList<DepthPoint>();
+    for (BigDecimal price : pricePoints) {
+      BigDecimal volume = buckets.get(price).getVolume();
+      result.add(new DepthPoint(price, volume));
+    }
+    return result;
+  }
 
-	public class DepthPoint {
-		final public Long price;
-		final public Long volume;
+  public class DepthPoint {
+    final public BigDecimal price;
+    final public BigDecimal volume;
 
-		public DepthPoint(Long price, Long volume) {
-			this.price = price;
-			this.volume = volume;
-		}
-	}
+    public DepthPoint(BigDecimal price, BigDecimal volume) {
+      this.price = price;
+      this.volume = volume;
+    }
+  }
 
-	@Override
-	public String toString() {
-		return new Gson().toJson(this);
-	}
+  @Override
+  public String toString() {
+    return new Gson().toJson(this);
+  }
 
 }
